cleaned up server.js
